Restructure the build script
