+++ .babelrc
@@ -1,0 +1,7 @@
+{
+  "presets": ["es2015"],
+  "plugins": [
+    "syntax-flow",
+    "transform-flow-strip-types"
+  ]
+}
