+++ package.json
@@ -4,10 +4,10 @@
   "description": "Website of 404 busters",
   "scripts": {
     "build": "NODE_ENV=production ./node_modules/.bin/gulp build",
-    "dev": "./node_modules/.bin/gulp dev",
-    "help": "./node_modules/.bin/gulp help",
-    "watch": "./node_modules/.bin/gulp watch",
-    "test": "echo 'No Test here'"
+    "dev": "gulp dev",
+    "help": "gulp help",
+    "watch": "gulp watch",
+    "test": "flow check"
   },
   "repository": {
     "url": "git+https://github.com/404busters/404busters.github.io",
@@ -15,15 +15,20 @@
   },
   "license": "MIT",
   "devDependencies": {
+    "autoprefixer": "^6.6.1",
     "babel-core": "^6.20.0",
     "babel-loader": "^6.2.9",
+    "babel-plugin-syntax-flow": "^6.18.0",
+    "babel-plugin-transform-flow-strip-types": "^6.21.0",
     "babel-preset-es2015": "^6.18.0",
+    "flow-bin": "^0.37.4",
     "gulp": "^3.9.1",
     "gulp-clean-css": "^2.2.2",
     "gulp-help": "^1.6.1",
     "gulp-htmlmin": "^3.0.0",
     "gulp-imagemin": "^3.1.0",
     "gulp-nunjucks": "^2.3.0",
+    "gulp-postcss": "^6.2.0",
     "gulp-rename": "^1.2.2",
     "gulp-sass": "^3.0.0",
     "gulp-util": "^3.0.7",
@@ -31,11 +36,5 @@
     "webpack": "^1.14.0",
     "webpack-dev-server": "^1.16.2",
     "webpack-merge": "^1.1.1"
-  },
-  "dependencies": {},
-  "babel": {
-    "presets": [
-      "es2015"
-    ]
   }
 }
